+++ asgardpy/data/dl4.py
@@ -10,7 +10,6 @@ from gammapy.estimators import FluxPointsEstimator
 from gammapy.modeling import Fit
 
 from asgardpy.base import AnalysisStepBase, BaseConfig, EnergyRangeConfig
-from asgardpy.data.target import get_chi2_pval
 
 __all__ = [
     "FitAnalysisStep",
@@ -64,19 +63,7 @@ class FitAnalysisStep(AnalysisStepBase):
 
         self.instrument_spectral_info["total_stat"] = self.fit_result.total_stat
 
-        chi2_, pval_ = get_chi2_pval(
-            self.instrument_spectral_info["total_stat"], self.instrument_spectral_info["DoF"]
-        )
-        self.instrument_spectral_info["chi2"] = chi2_
-        self.instrument_spectral_info["p-value"] = pval_
-
-        stat_message = (
-            f'The Chi2/dof value of the Fit is {chi2_:.2f}/{self.instrument_spectral_info["DoF"]}'
-        )
-        stat_message += f"\nand the p-value is {pval_:.3e}"
-
         self.log.info(self.fit_result)
-        self.log.info(stat_message)
 
     def _setup_fit(self):
         """